+++ implementation/js/src/infrastructure/lib+-/marionette/template-cache.js
@@ -3,18 +3,21 @@
 	//1 Override the default raw-template retrieving method
 	//We allow both #id or @*.html(remote) and template html string(or string array) as parameter.
 	Backbone.Marionette.TemplateCache.prototype.loadTemplate = function(idOrTplString){
+		//local tpl
 		if(_.string.startsWith(idOrTplString, '#')) return $(idOrTplString).html();
+		//remote tpl (with local stored map cache)
 		if(_.string.startsWith(idOrTplString, '@')) {
 			var name = idOrTplString.substr(1);
 			//search the local templates cache:
-			var tpl = app.Util.Tpl.get(name);
+			var tpl = app.Util.Tpl.remote.get(name);
 			if(tpl) return tpl;
-			//fetch from remote: (without CORS)
-			return app.Util.Tpl.load(app.config.viewTemplates + '/' + name, false, name).string;
+			//fetch from remote: (might need server-side CORS support)
+			return app.Util.Tpl.remote.load(name);
 
 		}
-		if(_.isArray(idOrTplString)) return idOrTplString.join('');
-		return idOrTplString; //this can NOT be null or empty since Marionette.Render guards it so don't need to use idOrTplString || ' ';
+		//string and string array
+		return app.Util.Tpl.build(idOrTplString);
+		//this can NOT be null or empty since Marionette.Render guards it so don't need to use idOrTplString || ' ';
 	};
 
 	//2 Override to use Handlebars templating engine with Backbone.Marionette